requirements.txt
asgiref==3.8.1
Django==4.2.18
django-debug-toolbar==5.0.1
django-environ==0.12.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-environ==0.4.54
sqlparse==0.5.3
typing_extensions==4.12.2
